Un-nest String.RuneEnumerator
